+++ Models/Api/ExerciseDone/ExerciseDoneFormModel.cs
@@ -2,9 +2,9 @@
 {
     public class ExerciseDoneFormModel
     {
-        public long ExerciseId { get; set; }
-        public long Reps { get; set; }
-        public long Weight { get; set; }
+        public int ExerciseId { get; set; }
+        public int Reps { get; set; }
+        public double Weight { get; set; }
         public string Equipment { get; set; }
         public int Pulse { get; set; }
     }
