+++ Models/Api/Training/TrainingFormModel.cs
@@ -4,9 +4,9 @@ namespace GymApp.Models.Api.Training
 {
     public class TrainingFormModel
     {
-        public long TrainingPlanId { get; set; }
+        public int TrainingPlanId { get; set; }
         public string Attention { get; set; }
-        public int AvgPulse { get; set; }
+        public double AvgPulse { get; set; }
         
         public List<ExerciseDoneFormModel> ExercisesDone { get; set; }
     }
